Fix: Widening conversion from Long to Double is deprecated
deprecated since scala 2.13.2

+++ akka-remote/src/main/scala/akka/remote/PhiAccrualFailureDetector.scala
@@ -204,7 +204,7 @@ class PhiAccrualFailureDetector(
       -math.log10(1.0 - 1.0 / (1.0 + e))
   }
 
-  private val minStdDeviationMillis = minStdDeviation.toMillis
+  private val minStdDeviationMillis = minStdDeviation.toMillis.toDouble
 
   private def ensureValidStdDeviation(stdDeviation: Double): Double = math.max(stdDeviation, minStdDeviationMillis)
